refactor: Separate price helper lps
