+++ src/config/constants/priceHelperLps/farms/1.ts
@@ -1,5 +1,5 @@
 // import { serializeTokens } from 'utils/serializeTokens'
-import { SerializedFarmConfig } from '../types'
+import { SerializedFarmConfig } from '../../types'
 // import { ethereumTokens } from '../tokens'
 
 // const serializedTokens = serializeTokens(ethereumTokens)
