+++ src/config/constants/priceHelperLps/farms/5.ts
@@ -1,5 +1,5 @@
 import { goerliTestnetTokens, serializeToken } from '@pancakeswap/tokens'
-import { SerializedFarmConfig } from '../types'
+import { SerializedFarmConfig } from '../../types'
 
 const priceHelperLps: SerializedFarmConfig[] = [
   {
