+++ src/config/constants/priceHelperLps/farms/56.ts
@@ -1,0 +1,5 @@
+import { SerializedFarmConfig } from '../../types'
+
+const priceHelperLps: SerializedFarmConfig[] = []
+
+export default priceHelperLps
