+++ src/config/constants/priceHelperLps/farms/97.ts
@@ -1,5 +1,5 @@
 // import { bscTestnetTokens } from '@pancakeswap/tokens'
-import { SerializedFarmConfig } from '../types'
+import { SerializedFarmConfig } from '../../types'
 
 const priceHelperLps: SerializedFarmConfig[] = []
