+++ src/config/constants/priceHelperLps/pools/5.ts
@@ -1,0 +1,5 @@
+import { SerializedFarmConfig } from '../../types'
+
+const priceHelperLps: SerializedFarmConfig[] = []
+
+export default priceHelperLps
