+++ src/config/constants/priceHelperLps/pools/56.ts
@@ -1,5 +1,5 @@
 import { bscTokens, serializeToken } from '@pancakeswap/tokens'
-import { SerializedFarmConfig } from '../types'
+import { SerializedFarmConfig } from '../../types'
 
 const priceHelperLps: SerializedFarmConfig[] = [
   /**
@@ -15,13 +15,6 @@ const priceHelperLps: SerializedFarmConfig[] = [
     lpAddress: '0x3147F98B8f9C53Acdf8F16332eaD12B592a1a4ae',
     token: bscTokens.ankr,
     quoteToken: bscTokens.wbnb,
-  },
-  {
-    pid: null,
-    lpSymbol: 'ANTEX-BUSD LP',
-    lpAddress: '0x4DcB7b3b0E8914DC0e6D366521604cD23E7991E1',
-    token: bscTokens.antex,
-    quoteToken: bscTokens.busd,
   },
 ].map((p) => ({ ...p, token: serializeToken(p.token), quoteToken: serializeToken(p.quoteToken) }))
