+++ src/config/constants/priceHelperLps/pools/97.ts
@@ -1,0 +1,6 @@
+// import { bscTestnetTokens } from '../tokens'
+import { SerializedFarmConfig } from '../../types'
+
+const priceHelperLps: SerializedFarmConfig[] = []
+
+export default priceHelperLps
