+++ src/state/farms/index.ts
@@ -14,7 +14,7 @@ import masterchefABI from 'config/abi/masterchef.json'
 import { getMasterChefAddress } from 'utils/addressHelpers'
 import { getBalanceAmount } from 'utils/formatBalance'
 import type { AppState } from 'state'
-import { getPriceHelperLpFiles } from 'config/constants/priceHelperLps'
+import { getFarmsPriceHelperLpFiles } from 'config/constants/priceHelperLps'
 import splitProxyFarms from 'views/Farms/components/YieldBooster/helpers/splitProxyFarms'
 import { getFarmConfig } from 'config/constants/farms/index'
 import fetchFarms from './fetchFarms'
@@ -80,7 +80,7 @@ export const fetchFarmsPublicDataAsync = createAsyncThunk<
     const farmsCanFetch = farmsConfig.filter(
       (farmConfig) => pids.includes(farmConfig.pid) && poolLengthAsBigNumber.gt(farmConfig.pid),
     )
-    const priceHelperLpsConfig = getPriceHelperLpFiles(chainId)
+    const priceHelperLpsConfig = getFarmsPriceHelperLpFiles(chainId)
 
     const farms = await fetchFarms(farmsCanFetch.concat(priceHelperLpsConfig), chainId)
     const farmsWithPrices = farms.length > 0 ? getFarmsPrices(farms, chainId) : []
